BRR BRR PATAPIN

+++ scripts/alterna_animacao.py
@@ -2,26 +2,31 @@ import os
 
 readme_path = "README.md"
 
-with open(readme_path, "r", encoding="utf-8") as file:
-    content = file.read()
+# Verifica se o arquivo README.md existe
+if os.path.exists(readme_path):
+    with open(readme_path, "r", encoding="utf-8") as file:
+        content = file.read()
 
-pacman_block = '''<div id="pacman">
-<picture>
-     <source media="(prefers-color-scheme: dark)" srcset="https://raw.githubusercontent.com/miguel-lamazares/miguel-lamazares/output/pacman-contribution-graph-dark.svg">
-     <source media="(prefers-color-scheme: light)" srcset="https://raw.githubusercontent.com/miguel-lamazares/miguel-lamazares/output/pacman-contribution-graph.svg">
-    <img alt="pacman contribution graph" src="https://raw.githubusercontent.com/miguel-lamazares/miguel-lamazares/output/pacman-contribution-graph.svg">
-</picture>
-</div>'''
+    pacman_block = '''<div id="pacman">
+    <picture>
+         <source media="(prefers-color-scheme: dark)" srcset="https://raw.githubusercontent.com/miguel-lamazares/miguel-lamazares/output/pacman-contribution-graph-dark.svg">
+         <source media="(prefers-color-scheme: light)" srcset="https://raw.githubusercontent.com/miguel-lamazares/miguel-lamazares/output/pacman-contribution-graph.svg">
+        <img alt="pacman contribution graph" src="https://raw.githubusercontent.com/miguel-lamazares/miguel-lamazares/output/pacman-contribution-graph.svg">
+    </picture>
+    </div>'''
 
-snake_block = '''<div id="snake" align="center">
- <img src="https://raw.githubusercontent.com/miguel-lamazares/miguel-lamazares/output/snake.svg" alt="Snake animation" />
-</div>'''
+    snake_block = '''<div id="snake" align="center">
+     <img src="https://raw.githubusercontent.com/miguel-lamazares/miguel-lamazares/output/snake.svg" alt="Snake animation" />
+    </div>'''
 
-# Alternar
-if pacman_block in content:
-    content = content.replace(pacman_block, snake_block)
-elif snake_block in content:
-    content = content.replace(snake_block, pacman_block)
+    # Alternar entre pacman e snake
+    if pacman_block in content:
+        content = content.replace(pacman_block, snake_block)
+    elif snake_block in content:
+        content = content.replace(snake_block, pacman_block)
 
-with open(readme_path, "w", encoding="utf-8") as file:
-    file.write(content)
+    # Atualiza o README.md
+    with open(readme_path, "w", encoding="utf-8") as file:
+        file.write(content)
+else:
+    print(f"O arquivo {readme_path} não foi encontrado.")
